+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This release adds the following features:
 
 - **Repeatable Auctions**: Governance can now specify the number of times an auction can be repeated while approving the auctions. This allows the auctions to be started up to the specified number of times until the other parameters are fulfilled, and the lot available in the auction is exhausted.
 
-- **Dust Limits**: Governance can now specify the minimum amount of tokens that would be considered "valuable" for the folio and disallow removing those tokens from the basket for any reason. These dust limits are also used to limit actions in auctions and other places where amount of tokens are important. Both the admin and auction approvers can set these limits.
+- **Dust Limits**: Governance can now specify the minimum amount of tokens that would be considered "valuable" for the folio and disallow removing those tokens from the basket for any reason. These dust limits are also used to limit actions in auctions and other places where amount of tokens are important. Both the admin and rebalance managers can set these limits.
 
 - **Minimum Mint Enforcement**: While minting Folio tokens, you can now specify a minimum amount out such that any changes to fees between you sending the transaction and inclusion would cause it to revert. The output amount must also be non-zero going forward.